Separated logic from the VM to ChartDataManager to save and create user score records for the insights Pie Charts.
